% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2.1_Rbc1.R
\name{is.done.Rbc1}
\alias{is.done.Rbc1}
\title{Check if Rbc1 procedure has completed}
\usage{
\method{is.done}{Rbc1}(x)
}
\arguments{
\item{x}{procedure object of class Rbc1}
}
\value{
Boolean
}
\description{
Check if Rbc1 procedure has completed
}
\examples{
proc <- Rbc1(2, 56, 8)
is.done(proc)
# FALSE
}

