% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/1_generic_functions.R
\name{is.done}
\alias{is.done}
\title{Generic function to determin whether or not a procedure has finished}
\usage{
is.done(proc, ...)
}
\arguments{
\item{proc}{S3 Class object for a specific selection procedure (\code{Rbc1}, \code{Rbc2}, \code{Rhbc1} or \code{Rhbc2})}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Generic function to determin whether or not a procedure has finished
}

