% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2.3_Rbc.helpers.R
\name{Power}
\alias{Power}
\title{Helper function to compute the power(\eqn{1-\beta}) of a Rbc1 procedure}
\usage{
Power(theta0, k, c, n, d0 = 0, d1 = 0.05, d2 = 0.2)
}
\arguments{
\item{theta0}{control Bernoulii parameter}

\item{k}{number of treatments}

\item{c}{design constant c}

\item{n}{number of maximum observations from each treatment, either a integer of a vector of intergerst with length k+1}

\item{d0}{design constant \eqn{\delta_0}, \eqn{\theta \leq \theta_0 + \delta_0} is considered ineffective}

\item{d1}{design constant \eqn{\delta_1}, \eqn{\theta \leq \theta_0 + \delta_1} is considered at most marginally effective}

\item{d2}{design constant \eqn{\delta_2}, \eqn{\theta \geq \theta_0 + \delta_1} is considered effective}
}
\description{
Helper function to compute the power(\eqn{1-\beta}) of a Rbc1 procedure
}
\examples{
power(0.2, 3, 8, 60) # roughly 0.75
}

