% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2.1_Rbc1.R
\name{Rbc1}
\alias{Rbc1}
\title{S3 Class constructor for Rbc1 procedure}
\arguments{
\item{k}{number of treatments}

\item{n}{number of maximum observations to take from each treatment}

\item{c}{design parameter c, cutoff point for choosing a treatment over control}
}
\value{
S3 Class object of the same procedure type
}
\description{
S3 Class constructor for Rbc1 procedure
}
\examples{
# run on a full dataset
# data is a data.frame of dimention (128 ,4), with 0/1 values

proc <- Rbc1(3, 13, 128, data)

# alternative
proc <- Rbc1(3, 13, 128)


summary(run.experiment(proc, data))
# Simulation Example
# Choose: k=3, P1=0.95, P2=0.95, theta_0 = 0.2, delta_1 = 0.05, delta_2 = 0.2
# From Table 2: we have, n = 128, c=13
n <- 128
k <- 3
proc <- Rbc1(k=k, c=13, n=n)
repeat {
  p <- c(0.2, runif(k))         # population bernoulie parameters
  jmax <- which.max(p[-1]) + 1  # best treatment
  if (p[jmax] >= p[1] + 0.2) {  # ensure assumptions in Section (2) are satisfied AND there's a best treatment to select
    invalid <- sapply(p[-1], function(x) {
      x > p[1] + 0.05 & x < p[1] + 0.2
    })
    if(all(!invalid)) break
  }
}
step <- 1
repeat {
  sampleObs <- sapply(p, function(prob) { # generate B(1, p) binary r.v. as observations, from population probs.
    rbinom(1, 1, prob)
  })
  proc <- vector.at.a.time(proc, sampleObs)
  step <- step + 1
  if (step > n | ( is.done(proc) )) break
}
summary(proc)
}

