% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/1_generic_functions.R
\name{run.experiment}
\alias{run.experiment}
\title{Generic function to run procedure with entire data set instead of sequentially (a.k.a. vector.at.a.time)}
\usage{
run.experiment(proc, data, ...)
}
\arguments{
\item{proc}{S3 Class object for a specific selection procedure (\code{Rbc1}, \code{Rbc2}, \code{Rhbc1} or \code{Rhbc2})}

\item{data}{data.frame object}
}
\value{
S3 Class object of the same procedure type
}
\description{
Generic function to run procedure with entire data set instead of sequentially (a.k.a. vector.at.a.time)
}

