% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2.3_Rbc.helpers.R
\name{Size}
\alias{Size}
\title{Helper function to compute the size(\eqn{\alpha}) of a Rbc1 procedure}
\usage{
Size(theta0, k, c, n, d0 = 0)
}
\arguments{
\item{theta0}{control Bernoulii parameter}

\item{k}{number of treatments}

\item{c}{design constant c}

\item{n}{number of maximum observations from each treatment, either a integer of a vector of intergerst with length k+1}

\item{d0}{design constant \eqn{\delta_0}, \eqn{\theta \leq \theta_0 + \delta_0} is considered ineffective}
}
\description{
Helper function to compute the size(\eqn{\alpha}) of a Rbc1 procedure
}
\examples{
size(0.2, 3, 8, 60) # roughly 0.05
}

