% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/1_generic_functions.R
\name{stage.one}
\alias{stage.one}
\title{Generic function to run stage one of two stage procedures}
\usage{
stage.one(proc, ...)
}
\arguments{
\item{proc}{S3 Class object for a specific selection procedure (\code{Rbc1}, \code{Rbc2}, \code{Rhbc1} or \code{Rhbc2})}

\item{\code{...}}{addtional arguments}
}
\value{
S3 Class object of the same procedure type
}
\description{
Generic function to run stage one of two stage procedures
}

