% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2.1_Rbc1.R
\name{summary.Rbc1}
\alias{summary.Rbc1}
\title{Summary of Rbc1 procedure}
\usage{
\method{summary}{Rbc1}(x)
}
\arguments{
\item{x}{procedure object of class Rbc1}
}
\value{
Table
}
\description{
Summary of Rbc1 procedure
}
\examples{
# Sample output
Procedure Rbc1, see Buzaianu and Chen (2008)
Procedure completed: select Treatement 2
At Step:  128
Control Treatment 1 Treatment 2 Treatment 3
No. Obs       128.0000   128.00000 128.0000000   128.00000
No. Successes  24.0000    12.00000  93.0000000    76.00000
p.esti.         0.1875     0.09375   0.7265625     0.59375
}

