% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2.2_Rbc2.R
\name{summary.Rbc2}
\alias{summary.Rbc2}
\title{Summary of Rbc2 procedure}
\usage{
\method{summary}{Rbc2}(x)
}
\arguments{
\item{x}{procedure object of class Rbc2}
}
\value{
Table
}
\description{
Summary of Rbc2 procedure
}
\examples{
# Sample output
Procedure Rbc2, see Buzaianu and Chen (2008)
Procedure completed: select Treatement 3
Used 365 observations in total, saving: 28.71\%
At Step:  97
Control Treatment 1 Treatment 2 Treatment 3
No. Obs            91          97          80          97
No. Successes      21          45          15          76
Stopped             1           1           1           1
}

