% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2.2_Rbc2.R
\name{vector.at.a.time.Rbc2}
\alias{vector.at.a.time.Rbc2}
\title{Vector-at-a-time sampling for Rbc2 procedure}
\usage{
\method{vector.at.a.time}{Rbc2}(x, vec)
}
\arguments{
\item{x}{procedure object of class Rbc2}

\item{vec}{an observation vector or 0/1 (alternative: TRUE/FALSE) of size k + 1}
}
\value{
S3 Class object of the same procedure type (Rbc2)
}
\description{
Vector-at-a-time sampling for Rbc2 procedure
}
\examples{
proc <- Rbc2(2, 56, 8)
# proc$k == 2
# is.done(proc) == FALSE
proc <- vector.at.a.time(proc, c(0,1,1))
summary(proc)
}

