% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/1_generic_functions.R
\name{vector.at.a.time}
\alias{vector.at.a.time}
\title{Vector-at-a-time sampling for procedures in this package}
\usage{
vector.at.a.time(proc, vec, ...)
}
\arguments{
\item{proc}{S3 Class object for a specific selection procedure (\code{Rbc1}, \code{Rbc2}, \code{Rhbc1} or \code{Rhbc2})}

\item{vec}{an observation vector or 0/1 (alternative: TRUE/FALSE) of size k + 1}
}
\value{
S3 Class object of the same procedure type
}
\description{
Vector-at-a-time sampling for procedures in this package
}

