% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/0_utils.R
\name{which.is.max}
\alias{which.is.max}
\title{Where is the max()}
\usage{
which.is.max(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
index of max value, break ties randomly
}
\description{
Determines the location, i.e., index of the (first) maximum of a numeric vector, break ties randomly; borrowed from nnet package
}

